+++ src/AngleSharp.Css/Parser/CssParserExtensions.cs
@@ -17,10 +17,7 @@ namespace AngleSharp.Css.Parser
         /// <param name="parser">The parser to extend.</param>
         /// <param name="content">The string content to parse.</param>
         /// <returns>The task yielding the created stylesheet.</returns>
-        public static Task<ICssStyleSheet> ParseStyleSheetAsync(this ICssParser parser, String content)
-        {
-            return parser.ParseStyleSheetAsync(content, CancellationToken.None);
-        }
+        public static Task<ICssStyleSheet> ParseStyleSheetAsync(this ICssParser parser, String content) => parser.ParseStyleSheetAsync(content, CancellationToken.None);
 
         /// <summary>
         /// Parses a CSS stylesheet from a string asynchronously.
@@ -28,10 +25,7 @@ namespace AngleSharp.Css.Parser
         /// <param name="parser">The parser to extend.</param>
         /// <param name="content">The stream content to parse asynchronously.</param>
         /// <returns>The task yielding the created stylesheet.</returns>
-        public static Task<ICssStyleSheet> ParseStyleSheetAsync(this ICssParser parser, Stream content)
-        {
-            return parser.ParseStyleSheetAsync(content, CancellationToken.None);
-        }
+        public static Task<ICssStyleSheet> ParseStyleSheetAsync(this ICssParser parser, Stream content) => parser.ParseStyleSheetAsync(content, CancellationToken.None);
 
         /// <summary>
         /// Parses a CSS stylesheet from hosting stylesheet asynchronously.
@@ -39,9 +33,6 @@ namespace AngleSharp.Css.Parser
         /// <param name="parser">The parser to extend.</param>
         /// <param name="sheet">The stylesheet containg the URL reference.</param>
         /// <returns>The task yielding the finished stylesheet.</returns>
-        public static Task<ICssStyleSheet> ParseStyleSheetAsync(this ICssParser parser, ICssStyleSheet sheet)
-        {
-            return parser.ParseStyleSheetAsync(sheet, CancellationToken.None);
-        }
+        public static Task<ICssStyleSheet> ParseStyleSheetAsync(this ICssParser parser, ICssStyleSheet sheet) => parser.ParseStyleSheetAsync(sheet, CancellationToken.None);
     }
 }